Cleaning up and organizing the list screens since now the lists and details are handled in one screen
