+++ gradle/libs.versions.toml
@@ -26,7 +26,7 @@ ktorVersion = "2.3.6"
 tvCompose = "1.0.0-alpha10"
 workVersion = "2.9.0"
 ziplineVersion = "1.1.0"
-landscapist = "2.2.11"
+landscapist = "2.2.12"
 protobufVersion = "3.25.1"
 sketchVersion = "3.2.5"
 activity = "1.8.1"
@@ -87,9 +87,10 @@ composeRuntimeLivedata = { group = "androidx.compose.runtime", name = "runtime-l
 composeLifecycle = { module = "androidx.lifecycle:lifecycle-viewmodel-compose", version.ref = "lifecycle" }
 composeActivity = { module = "androidx.activity:activity-compose", version.ref = "activity" }
 composeLifecycleRuntime = { module = "androidx.lifecycle:lifecycle-runtime-compose", version.ref = "lifecycle" }
-landscapistGlide = { module = "com.github.skydoves:landscapist-glide", version.ref = "landscapist" }
-landscapistPalette = { module = "com.github.skydoves:landscapist-palette", version.ref = "landscapist" }
-landscapistPlaceholder = { module = "com.github.skydoves:landscapist-placeholder", version.ref = "landscapist" }
+landscapist-bom = { module = "com.github.skydoves:landscapist-bom", version.ref = "landscapist" }
+landscapistGlide = { group = "com.github.skydoves", name = "landscapist-glide" }
+landscapistPalette = { group = "com.github.skydoves", name = "landscapist-palette" }
+landscapistPlaceholder = { group = "com.github.skydoves", name = "landscapist-placeholder" }
 composeConstraintLayout = "androidx.constraintlayout:constraintlayout-compose:1.0.1"
 coilCompose = { module = "io.coil-kt:coil-compose", version.ref = "coil" }
 coilVideo = { module = "io.coil-kt:coil-video", version.ref = "coil" }
@@ -292,7 +293,7 @@ compose = [
     "composeAnimation",
     "composeActivity", "composeLifecycle", "composeLifecycleRuntime",
     "composeRuntimeLivedata",
-    "landscapistGlide", "landscapistPalette", "landscapistPlaceholder",
+    "landscapist-bom", "landscapistGlide", "landscapistPalette", "landscapistPlaceholder",
     "coilCompose",
     "composeConstraintLayout", "permissions",
     "drawablePainter", "uiUtil",
@@ -305,7 +306,7 @@ composeTv = [
     "composeAnimation",
     "composeActivity", "composeLifecycle", "composeLifecycleRuntime",
     "composeRuntimeLivedata",
-    "landscapistGlide", "landscapistPalette", "landscapistPlaceholder",
+    "landscapist-bom", "landscapistGlide", "landscapistPalette", "landscapistPlaceholder",
     "coilCompose",
     "composeConstraintLayout", "permissions",
     "drawablePainter", "uiUtil",
